validating chain
