+++ tests/MQTTnet.Extensions.MultiCloud.UnitTests/ca.pem
@@ -1,0 +1,32 @@
+-----BEGIN CERTIFICATE-----
+MIIFkTCCA3mgAwIBAgIUfmwKHpXyiSBf3Shznv9pb7SvQbgwDQYJKoZIhvcNAQEL
+BQAwWDEVMBMGA1UEAwwMUmlkbyBGWTIzIENBMQswCQYDVQQGEwJEQzEPMA0GA1UE
+CAwGR290aGFtMQ8wDQYDVQQHDAZHb3RoYW0xEDAOBgNVBAoMB0JhdENhdmUwHhcN
+MjIwODMxMjExMjQ2WhcNMjcwODMxMjExMjQ2WjBYMRUwEwYDVQQDDAxSaWRvIEZZ
+MjMgQ0ExCzAJBgNVBAYTAkRDMQ8wDQYDVQQIDAZHb3RoYW0xDzANBgNVBAcMBkdv
+dGhhbTEQMA4GA1UECgwHQmF0Q2F2ZTCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCC
+AgoCggIBAP1wy1LrTkQnafpG3r1wr/zcAcBfD4MgRa/8qvctdW5UbDS5ycKBfW8e
+VeLPcA3d46JnE4hgGdDbbKEMetNjgYJ/W16Crec0sAn1Gde3T4UyarQ7eB+eSB3C
+zFo3r6fe3e1a1Lub6R8DuZLD1QXXgfc1c5eKsdi0cZCKQXTj47KFfYRXbpIm7mIV
+17zY4CJvjqMUnOOYCeNOYfKcKu8DQui6zIoxa6Hpkl7wMydyheWbKGHofCDe91gU
+F+5+QAKy+UPpXY0Nv1pT9T+ilW1rFDqq07u/uBWVarplAYotUGMFD3e5GjUG1iVb
+fs9+lZpQtSx688bpPJMO1d7eNUlpjkhG1AYeLd2/SlyO6NQjmXCfZ67f8JTKhe0X
+phgG3VcFhR+LZ+JUIrBDpBEUOAaQuykDCKQKN/E2NrnzY+kE/pc/Gz1XDFrNX2p5
+jE+dwMufvFP6EorBaSaNHF4S6QFQMMrgyiGZEV1eVhtoz433iS7eyAxclEV9/qRG
+DwEVCtuABe7kkXE2C7hCuSLLE5KTAvoKqDX9MjDpj2ke4r3OJ2P/IjxsJHRffGfn
+DDbSNn932QsFbs7MltaUFK8/TOgmqulcfHktZtUWZXIQSNRzKIJ+RpQyaH2TBWMd
+g2oeAqELKoEubU5ecRbFqPilqs6fYE0AQoCsw6BkYScA2De3SxJtAgMBAAGjUzBR
+MB0GA1UdDgQWBBTDYWJ7DGKdmWvdSLoGjSCmfQOqhDAfBgNVHSMEGDAWgBTDYWJ7
+DGKdmWvdSLoGjSCmfQOqhDAPBgNVHRMBAf8EBTADAQH/MA0GCSqGSIb3DQEBCwUA
+A4ICAQDBHLDlOK2X9Jc1FazgUF/2Hh9G2PGYbpt5z8rMUQVAJsLBERNvAom0coRL
+ccBEfz8JaQOPNoaqejypTrSIoZXdYziFsS1IWF68v96vG0C6pIyzon9mjCvNIJU5
+0DQXSWw1B8lP/q6FUKndi2b55Qx0xVlhobRsSxt/DPIpOPc4wHdHxgeO6kHWa7CK
+/KmrjNU8FzZz2F6CnoR2CHA4BQ7y4aV44SatioAlIeJ7iOV+D+ZEpa+7lQdUNMeX
+WyOi9DN9tKH8ar84ykrA+IOBOUnV+/TENGNcxqycSUpKW54J1V5/gTze/AdI7w1O
+2YmkI0mjTPZdCz5s8Xinun666MXJ7fCUlc566/Twr9W26pcAc/LXbJBfZZqNMNYL
+Pzyzozuw/6isrqSfyxI9R9yM03KLD8T0rMlmGr4PYtSKZF2IJLKcmSL2s36hcy21
+5aLhbuKZ1dmHT4HP80OTcfbDv76r6Ufn1ndcz1uwuBSStvSX3klRPckeGGYV/lgf
+K3jtwgmrFTomo2aBt+Ivx4PwVMOfkwHui2WD2ksaRy+5LhhLH4Q+l86EaM7rCfuV
+h95BYHQGe52axKpot6idwly+IJa2H6tvG5qNqv+lMpnZTZnWL5vDSE20aZS8k/dG
+QZsrzCE+ZeisiVCMcN7rLpTzTLOTFtBcKuAP7rFffdZysSBvaA==
+-----END CERTIFICATE-----
