+++ tests/MQTTnet.Extensions.MultiCloud.UnitTests/caWithChain.pem
@@ -1,0 +1,78 @@
+-----BEGIN CERTIFICATE-----
+MIIGvTCCBKWgAwIBAgICA+kwDQYJKoZIhvcNAQELBQAwgZoxCzAJBgNVBAYTAlVT
+MRMwEQYDVQQIDApXYXNoaW5ndG9uMRAwDgYDVQQHDAdSZWRtb25kMSIwIAYDVQQK
+DBlEZWZhdWx0IEVkZ2UgT3JnYW5pemF0aW9uMSIwIAYDVQQLDBlEZWZhdWx0IEVk
+Z2UgT3JnYW5pemF0aW9uMRwwGgYDVQQDDBNFZGdlIFRlc3QgRGV2aWNlIENBMB4X
+DTIyMTEwMjIyMDA1OVoXDTIzMTEwMjIyMDA1OVowgZQxCzAJBgNVBAYTAlVTMRMw
+EQYDVQQIDApXYXNoaW5ndG9uMRAwDgYDVQQHDAdSZWRtb25kMSIwIAYDVQQKDBlE
+ZWZhdWx0IEVkZ2UgT3JnYW5pemF0aW9uMSIwIAYDVQQLDBlEZWZhdWx0IEVkZ2Ug
+T3JnYW5pemF0aW9uMRYwFAYDVQQDDA1FZGdlIEFnZW50IENBMIICIjANBgkqhkiG
+9w0BAQEFAAOCAg8AMIICCgKCAgEA2TBCdUflz0j5rgNvNxz+xgLIECFoR2RFjPPI
+3Zh+zRMcU+vcvch33TkPDKQd1W4bBQtNPoJ3yNA97GRmhJew2yYnish756va3nOL
+UompgXrKVrnsVXqac0zlp67RFT4hPJ78WWcplSgiKIh6SvfgJoONlJZAcxHVD6hF
+Sd1L5BfRACb83ejQusxjrtclfBt+wVl51i2xfqy9u+E9N04efSspbnd2slynf0Vb
+wVYSodzimpXZev7ONBTqMYiheCsEgumGcCyTl+OWajLf729HWFgBV/UOciYfVfIx
+y2vGVOHxPgc4uKmbe7v9JBTzxOumOJSTPxcVUbuaK/dxldbDAFpyM8QCDGE3XBk2
+2XpavWEENLoLB4ZsB5/4H2PypzWREgeZEZ0hhfFI1g+mepsXCXBXjgV3uwoNrP5Z
+DKvpHpSo33JjdjWUP1VFNJ1bNG/WiQ0RvOyZRShXZ1FueBM9gyge6X0nT+IyHY2n
+ac5ibrm1sLSp3nIdhWvfcRvoHiromexNGXTDFoS1fpYH17odLL8FNedQoOT4wvm9
+PvJknYQsqF7Fvk/Yfhna7c9v8my5hQJXEan5GgZ14GvPgyOPtLfEQaB6B0i8C+A9
+hw8w5FOxC5EQHOOlmBR7zhCKDaZ8Vb9DP/ok/Wu6fqjTNWLlJdWFQ0MAcc49j3iK
+nBZUQ/MCAwEAAaOCAQ8wggELMA8GA1UdEwEB/wQFMAMBAf8wHQYDVR0OBBYEFMZE
+Ozfv9uAZ6AiifQgC/kVLDYVvMA4GA1UdDwEB/wQEAwIBhjCByAYDVR0jBIHAMIG9
+gBTRx3zYyZlwNM92Sz9v1nk8g7KBlaGBoKSBnTCBmjELMAkGA1UEBhMCVVMxEzAR
+BgNVBAgMCldhc2hpbmd0b24xEDAOBgNVBAcMB1JlZG1vbmQxIjAgBgNVBAoMGURl
+ZmF1bHQgRWRnZSBPcmdhbml6YXRpb24xIjAgBgNVBAsMGURlZmF1bHQgRWRnZSBP
+cmdhbml6YXRpb24xHDAaBgNVBAMME0VkZ2UgVGVzdCBEZXZpY2UgQ0GCAgPoMA0G
+CSqGSIb3DQEBCwUAA4ICAQCCj+hfCC/acCdklv3kbMcz84+eejSmglg0t0+97M1k
+0+1aqgwrMPiC3L9NM1V/exKPGpiEYQPzXP71/cHnRaS0nAbTXsQRDYapcGxqS7Tr
+8IId8ntYXt5ymZ6XoIYX75ImgQRn5qfFNlTS6g5IRycwmBtlRpYN/OWZgLm48s73
+mcFCkFb0xDtQHI4lowRtSDBxtTXUVXuqg8QqIfFMkf3twE7bz3RIErIbKrYkrJmx
+7q+RyMxYdZmmjRwmZHq9naW0OgXuY+3D3ASiGM//KmBdaCiLxXaLEPlUKr+4S02h
+FuH+Au73/Ax6om7zJ/xh/ldTYHg4YhAlPTZLzKbCzVhO5NeC7lowBoIrDHv074PW
+FXjskrlxeJcuipmBDho5lBlfy6VGo7YNcyxpyV3ji07Mj511O9Tu2kHZdDB7I+pN
+zV3AguRlFmF/Z5i+ujMZBcnWcYyldOX2NO+qNdHV/WpASmdmdM+YlZsaEWi4CKog
+bthyzVd6DdsyVcDyj6MfY7MUEitzjdnjIYf/OCSdRSLkSW5tRaQxhjWcfA0KLfVv
+1kzGLTEllo7Fiapp5u2AsDm9ww5517UECUBDQxzM87bWYkYAQHzOaG8/7Jp31qgt
+n/LwCSQf0cOUzp3xDn1SVf3tqzMmxsgdMDBi+HauEgp+tOjKvlrh02MQLWZTw7PY
+rA==
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIIGwzCCBKugAwIBAgICA+gwDQYJKoZIhvcNAQELBQAwgZoxCzAJBgNVBAYTAlVT
+MRMwEQYDVQQIDApXYXNoaW5ndG9uMRAwDgYDVQQHDAdSZWRtb25kMSIwIAYDVQQK
+DBlEZWZhdWx0IEVkZ2UgT3JnYW5pemF0aW9uMSIwIAYDVQQLDBlEZWZhdWx0IEVk
+Z2UgT3JnYW5pemF0aW9uMRwwGgYDVQQDDBNFZGdlIFRlc3QgRGV2aWNlIENBMB4X
+DTIyMTEwMjIyMDA1NVoXDTIzMTEwMjIyMDA1NVowgZoxCzAJBgNVBAYTAlVTMRMw
+EQYDVQQIDApXYXNoaW5ndG9uMRAwDgYDVQQHDAdSZWRtb25kMSIwIAYDVQQKDBlE
+ZWZhdWx0IEVkZ2UgT3JnYW5pemF0aW9uMSIwIAYDVQQLDBlEZWZhdWx0IEVkZ2Ug
+T3JnYW5pemF0aW9uMRwwGgYDVQQDDBNFZGdlIFRlc3QgRGV2aWNlIENBMIICIjAN
+BgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEAqF7nSKmwlZzPbDE8RSNSDG0iG9lA
+VjBcZ+WOLIXRmie+h0nv88/PBfmVnr3FbW71l7ttt5rqfG20HhubIg8VtZ3Kcoo4
+xTJodUM1HpGB+SeTuqosyT+JTubQrMc3YUAYP3o1xJNneJBL9CLi86yfJROHHvUN
+kQMTs3my/Z6S4Te4xaV69nyDLeWfgsTCQXy31yyRBmA3DBzgnmEAmy7LDxVdnr3b
+zVwPB/n4SidT1tqgAgZ0li7uMk2WA1i4+7vQuRiipQwXOBlRa9vzj1c1gNCMi77L
+1sF+PyLe4+8NtNqbvTZAB3SRsRFsPNlj+5iQRaD0htJ1ATrNMNI5/apUNzUacfMn
+cj4FhriXixa8fZ3JmWTXah1mNxirXbh3g+/t0G7ljQjFvEN8uEJD1e6/MQgvee+6
+a9zrOWBS8Dd22SeMwIiqYjhz+uaAWioPF7p3L83Sk854QUKbBCKQIf3iCn/R46cr
+xUtcf1YfE6HGTe1SGiiSwlNNPWmI3eFaqF6zNCFc1I8p09rE8LfUX85NzA6d5mEz
+HMJIeLUF5sm72qHWRzvNiAqNRswE7Rt5TaGTevysONJbLzOhJuaIsDge04U6XsN/
+D2gg6KTMG4Eu3WGAbV9nzn3lOZeIzb3N6mT7DNi7iNIjSpeye+hAyiupSPS7gr+z
+QfUfpyywRlFM/30CAwEAAaOCAQ8wggELMA8GA1UdEwEB/wQFMAMBAf8wHQYDVR0O
+BBYEFNHHfNjJmXA0z3ZLP2/WeTyDsoGVMA4GA1UdDwEB/wQEAwIBhjCByAYDVR0j
+BIHAMIG9gBTRx3zYyZlwNM92Sz9v1nk8g7KBlaGBoKSBnTCBmjELMAkGA1UEBhMC
+VVMxEzARBgNVBAgMCldhc2hpbmd0b24xEDAOBgNVBAcMB1JlZG1vbmQxIjAgBgNV
+BAoMGURlZmF1bHQgRWRnZSBPcmdhbml6YXRpb24xIjAgBgNVBAsMGURlZmF1bHQg
+RWRnZSBPcmdhbml6YXRpb24xHDAaBgNVBAMME0VkZ2UgVGVzdCBEZXZpY2UgQ0GC
+AgPoMA0GCSqGSIb3DQEBCwUAA4ICAQBU00kZEPH36b6FlvCYp8SDk2UB6lJHnN2H
+nDcPFsJQvhrXe0aKPW0a7IwbKdHbgg9IcQIl4zAukW7hmjgXsPXiEZDM08zopdTl
+KITurqioIDFw15912Ix+cJI3dA1n0GAHV/+gbLFOgFI3Wr6+haO+FXErWyPTqsMk
+9aY7PGkZuzf2Z+X0lLpmFFj0KK5fXEW6OqNyE5oO1Q4ff/sR754lhs75R8Wn7Wxv
+Lejczpz2mCcKIXNa1MSv8vVMw4YKDoOs0EyNLm1JKesEr3tjrUUZ9F+pDPxTJ3q3
+3mU133/Pch7ApdQu3nIPV6oYQ5fKVBcDBlOUqOezvsaj0fXVgIki/qvA7uTxRQdR
+cC8Ye1aYcJOnulMycAck04zLB9v0ILWhw6iF9PQ1J54kcUmxeKBMBdKb6KUndtu5
+tnpMPcbikPVd6PC+cV8EbbBiAw6ROYQUr6IJbFdoTvPG+W+rgQyK3ZS6Q32/Wuhd
+eWVa9527ZWFt5UEwA3mC/j1ztCVUL6jx7LHWUbGrZekPf+7qOeiNIb5Evbp1Rdvl
+D70y23XRdpJwlPmqr2kuAZS3jOUUy0zwznKVF+cxxhTOf5G/numw236h39K+nzzD
+alcgO9djPt2VnQEqK+Wr2OD3sHRjVM5JdeHGBspcbPBI32EwUYYxmeBFDWUM+6nR
+iNohCm+jEA==
+-----END CERTIFICATE-----
