+++ tests/MQTTnet.Extensions.MultiCloud.UnitTests/onething.pem
@@ -1,19 +1,26 @@
 -----BEGIN CERTIFICATE-----
-MIIDDzCCAfegAwIBAgIRAOai03CRCtdQt4l8n/0IL2kwDQYJKoZIhvcNAQELBQAw
-FTETMBEGA1UEAwwKcmlkb2NhZnkyMjAeFw0yMjA2MDcxOTUwNTJaFw0yMjA5MDUx
-OTUwNTJaMBMxETAPBgNVBAMTCG9uZXRoaW5nMIIBIjANBgkqhkiG9w0BAQEFAAOC
-AQ8AMIIBCgKCAQEAqeQD/D8vgkTjIhnOBW+EzCogMVtA/8jezOp4B1PflCRLV++1
-ZGskhiBa938p/6+LCEnbnAib0EaoQtiv/Dfy6zwkEiV/iS7ZyIsIrwDHF2CPS36H
-QFOOGE7kYuh+bPKmpok2e4D3kgPb7X2q/N5ShWBP+u5Y4VMrFuw4ieS+iewvuLmS
-BePJZwDXfgp54wi/1f9UqFfnZRtA8zQWEaVxGJAfC8X2yeBQ89NuGsa+kOAh1F9D
-6ZmUpahD5HVKwqXSl51hRRdmLbtkUnJvVcFR7IRZkwlFP1+h+sfqjLTb/rskl238
-82+Y/joBU7JtxU6c4p4LtL9qTx/vUj3E/TDRBQIDAQABo1wwWjAOBgNVHQ8BAf8E
-BAMCBaAwDAYDVR0TAQH/BAIwADA6BgNVHREEMzAxhi9zcGlmZmU6Ly9jbHVzdGVy
-LmxvY2FsL25zL2NlcnQtbWFuYWdlci9vbmV0aGluZzANBgkqhkiG9w0BAQsFAAOC
-AQEAcovD372wluuUFqRiy/kLgGWvl07365KrseWeOaHkzNcoYLOvxRNKPlU8eIfX
-yOCT2WCbYMp6z1NzyUyNl/hKcCLfxTNDJTbwiOPawm62zb+rUgi6IDxpmjgARj79
-p5l6jDitRgAaOtLYSjNAHzEw+acCeBSooykOyWw6WO2BlyLWw0lIJcwnZx1ZHnUJ
-8Ou2BSUGu5Hu7Adaxhv6N9Y5tj8WCddeojTnks9L8euzSK/yqhzbEkrNFr3VQfTP
-+aWda+qClyOotmeL7cSgYRApQU6agkZTvEFeloullvQm6lOp5r7ujMt4wYONjZwo
-BQBG8zLdMyGPZcnjIN9N2RcNSQ==
+MIIEZjCCAk6gAwIBAgIQGecn2dSR/YhIPLcETm9NrDANBgkqhkiG9w0BAQsFADBY
+MRUwEwYDVQQDDAxSaWRvIEZZMjMgQ0ExCzAJBgNVBAYTAkRDMQ8wDQYDVQQIDAZH
+b3RoYW0xDzANBgNVBAcMBkdvdGhhbTEQMA4GA1UECgwHQmF0Q2F2ZTAeFw0yMjEx
+MDkwMTIwNDJaFw0yNDExMDkwMTMwNDFaMBMxETAPBgNVBAMMCG9uZXRoaW5nMIIB
+IjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAvsQFDeazUfWIePZQsWjVrMHR
+rUPq2ciGJarhw9HmpyIr2lPGZHyJ5a++/lqIyW/Yrr3OTOyMl6e/aQ8A6nR6oI3F
+Dd4sPGUJ1Za6XjMLZYbJLlSHB9I5a0qTRWPgKLEQ7Kx2oi3Dx3o3IGbBKv5EWY9A
+yYQnfjGTvm044TDidazyS9WoHEugtSmif1/xqXnEPi12WWSf9zKHYnhxZYgFZpAP
+BJtIA6gVly2R69vzxb5iHN5rBpvR8mvrBF6E+aW+Ua+yT+VvOBRyaZFR/ywH6UlG
+mUnL+h8Svq87LIpkrXVZV2uDrKIv+/XFhEjxtJpAnOMIxTe9KVtxCu1mMMALyQID
+AQABo3EwbzAOBgNVHQ8BAf8EBAMCBPAwHQYDVR0lBBYwFAYIKwYBBQUHAwIGCCsG
+AQUFBwMBMB8GA1UdIwQYMBaAFMNhYnsMYp2Za91IugaNIKZ9A6qEMB0GA1UdDgQW
+BBSqotrqgYhasdvs5HvYCemIi7dKPDANBgkqhkiG9w0BAQsFAAOCAgEATjv6/f9o
+T5zzLwHEjK7DKvHz819c6wmdJM+kMqzdzy6sdHkYSfgyH77wml3Uz1G2OcH+R3Yq
+7cZ4YvKKtAsv/jf7HVsMPBikj3yUjIAYUtGnzyG94bnAX0iNDSttwoV9Cf7Fd8t/
+nPoPoNN6ISbC2WfJCpvYMXFdiBW+8XGEP0TrHXzjyDnqrwJIpx1HWQbGi5T1G7SP
+olmb3FMGDXp3e+26spOPMt5fyFYeTntLSmUdVg29jAJmSYG+g8+hbnsMEXBj8ZJY
+UBlQI8EBuKtLigfr6FnW8mOJNApLvMOncD/3eJz3B+zmqFf0kFkC6d0YxfD4nyPO
+ksLmMvYtxi8EfbuwOzLwy0321PfMi4EkwMbcGlZ4eVOmu4R2Fslae9+s2+U4ggaO
+I0lTdCZ9l82vzQPst2Ax7+sv+xaMXqegDu2P6N8bIEdrD983BguHsNICLCllIMI8
+Tq6yVVT0PVAJR2AVpEnvQ28MwW1M+aFnarWMJiFJiz07vkcyTVAOwlZai+htdylQ
+TXckzbq5GMiTaJIwWov9NtXHTyGSzHf98REDttgcwKN6uxUOUYoKfT5zrE3FpULR
+yf3VrZw+/QxkezfbetkKj+bQIDarDHcaRb+KnFXUsjFC05ROQA4JmfmNbcYQuj/m
+jNGhZAclvbRWTGAOvz2HsDQSnlHyHxfmNUc=
 -----END CERTIFICATE-----
